requirements.txt
django_nocaptcha_recaptcha==0.0.20
django_summernote==0.8.8.7
Django==2.0.7
django_smalluuid==1.1.1
django-bootstrap4==0.0.6
whitenoise==3.3.1
python-decouple==3.1
gunicorn==19.9.0